add init_with_layer to support additional use cases
